dashboard/requirements.txt
streamlit==1.28.0
plotly==5.18.0
pandas==2.1.0
snowflake-connector-python==3.7.0
python-dotenv==1.0.0